*fixed most -Wall warnings
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2718 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/Object.cpp
@@ -2276,7 +2276,7 @@ void Object::SpellNonMeleeDamageLog(Unit *pVictim, uint32 spellID, uint32 damage
 //==========================================================================================
 //==============================Post Damage Processing======================================
 //==========================================================================================
-	if (dmg.resisted_damage==dmg.full_damage && !abs_dmg)
+	if ((int32)dmg.resisted_damage==dmg.full_damage && !abs_dmg)
 	{		//Magic Absorption
 		if (pVictim->IsPlayer())
 		{
